+++ packages/core/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.8.0 — not yet released
 
 * The `NT` type parameter of `SingleRef<NT>` now extends the `Node` base type.
+* Move `IdOrUnresolved` type to the `class-core` package, as it was used only there and in `delta-protocol-common`.
+    * Remove the `serializedRef` function as it was a duplicate of `idFrom` in `class-core`.
 
 
 ## 0.7.2